Feat/delete dashboard

+++ src/modules/dashboards/saga.ts
@@ -1,4 +1,11 @@
-import { takeLatest, put, select, all, getContext } from 'redux-saga/effects';
+import {
+  takeLatest,
+  put,
+  select,
+  take,
+  all,
+  getContext,
+} from 'redux-saga/effects';
 import { push } from 'connected-react-router';
 
 import {
@@ -13,6 +20,8 @@ import {
   saveDashboard as saveDashboardAction,
   viewDashboard as viewDashboardAction,
   deleteDashboard as deleteDashboardAction,
+  showDeleteConfirmation,
+  hideDeleteConfirmation,
 } from './actions';
 
 import { serializeDashboard } from './serializers';
@@ -29,7 +38,7 @@ import {
   getWidgetSettings,
 } from '../widgets';
 
-import { BLOB_API, ROUTES } from '../../constants';
+import { BLOB_API, NOTIFICATION_MANAGER, ROUTES } from '../../constants';
 import {
   INITIALIZE_DASHBOARD_WIDGETS,
   FETCH_DASHBOARDS_LIST,
@@ -38,6 +47,8 @@ import {
   EDIT_DASHBOARD,
   DELETE_DASHBOARD,
   VIEW_DASHBOARD,
+  CONFIRM_DASHBOARD_DELETE,
+  HIDE_DELETE_CONFIRMATION,
 } from './constants';
 
 import { RootState } from '../../rootReducer';
@@ -111,14 +122,34 @@ export function* deleteDashboard({
   payload,
 }: ReturnType<typeof deleteDashboardAction>) {
   const { dashboardId } = payload;
+  yield put(showDeleteConfirmation(dashboardId));
+  const notificationManager = yield getContext(NOTIFICATION_MANAGER);
 
-  try {
-    const blobApi = yield getContext(BLOB_API);
-    yield blobApi.deleteDashboard(dashboardId);
+  const action = yield take([
+    CONFIRM_DASHBOARD_DELETE,
+    HIDE_DELETE_CONFIRMATION,
+  ]);
 
-    yield put(deregisterDashboard(dashboardId));
-  } catch (err) {
-    console.error(err);
+  if (action.type === CONFIRM_DASHBOARD_DELETE) {
+    yield put(hideDeleteConfirmation());
+    try {
+      const blobApi = yield getContext(BLOB_API);
+      yield blobApi.deleteDashboard(dashboardId);
+
+      yield put(deregisterDashboard(dashboardId));
+      yield notificationManager.showNotification({
+        type: 'info',
+        message: 'notifications.dashboard_delete_success',
+        autoDismiss: true,
+      });
+    } catch (err) {
+      yield notificationManager.showNotification({
+        type: 'error',
+        message: 'notifications.dashboard_delete_error',
+        showDismissButton: true,
+        autoDismiss: false,
+      });
+    }
   }
 }
